fix: structure refactor
