+++ cmd/devstream/verify.go
@@ -3,7 +3,6 @@ package main
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/devstream-io/devstream/internal/pkg/configloader"
 	"github.com/devstream-io/devstream/internal/pkg/pluginengine"
 	"github.com/devstream-io/devstream/pkg/util/log"
 )
@@ -17,14 +16,7 @@ var verifyCMD = &cobra.Command{
 
 func verifyCMDFunc(cmd *cobra.Command, args []string) {
 	log.Info("Verify started.")
-
-	gConfig, err := configloader.LoadGeneralConf(configFile)
-	if err != nil {
-		log.Errorf("Error: %s.", err)
-		return
-	}
-
-	if pluginengine.Verify(gConfig.ToolFile, gConfig.VarFile) {
+	if pluginengine.Verify(configFile) {
 		log.Success("Verify succeeded.")
 	} else {
 		log.Info("Verify finished.")
